+++ tracer/src/Datadog.Trace/Ci/TestOptimization.cs
@@ -18,6 +18,7 @@ using Datadog.Trace.Logging;
 using Datadog.Trace.Pdb;
 using Datadog.Trace.PlatformHelpers;
 using Datadog.Trace.Telemetry;
+using Datadog.Trace.Telemetry.Metrics;
 using Datadog.Trace.Util;
 using TaskExtensions = Datadog.Trace.ExtensionMethods.TaskExtensions;
 
@@ -625,6 +626,7 @@ internal sealed class TestOptimization : ITestOptimization
                 }
 
                 FlakyRetryFeature = TestOptimizationFlakyRetryFeature.Create(settings, remoteSettings);
+                RecordDynamicAtrTelemetry(settings);
                 DynamicInstrumentationFeature = TestOptimizationDynamicInstrumentationFeature.Create(settings, remoteSettings);
                 KnownTestsFeature = TestOptimizationKnownTestsFeature.Create(settings, remoteSettings, client);
                 EarlyFlakeDetectionFeature = TestOptimizationEarlyFlakeDetectionFeature.Create(settings, remoteSettings, KnownTestsFeature);
@@ -677,11 +679,23 @@ internal sealed class TestOptimization : ITestOptimization
         using var cd = CodeDurationRef.Create();
         var remoteSettings = TestOptimizationClient.CreateSettingsResponseFromTestOptimizationSettings(settings, tracerManagement);
         FlakyRetryFeature = TestOptimizationFlakyRetryFeature.Create(settings, remoteSettings);
+        RecordDynamicAtrTelemetry(settings);
         DynamicInstrumentationFeature = TestOptimizationDynamicInstrumentationFeature.Create(settings, remoteSettings);
         KnownTestsFeature = TestOptimizationKnownTestsFeature.Create(settings, remoteSettings, client);
         EarlyFlakeDetectionFeature = TestOptimizationEarlyFlakeDetectionFeature.Create(settings, remoteSettings, KnownTestsFeature);
         ImpactedTestsDetectionFeature = TestOptimizationImpactedTestsDetectionFeature.Create(settings, remoteSettings, environmentValues);
         SkippableFeature = TestOptimizationSkippableFeature.Create(settings, remoteSettings, client, this);
         TestManagementFeature = TestOptimizationTestManagementFeature.Create(settings, remoteSettings, client);
+    }
+
+    private static void RecordDynamicAtrTelemetry(TestOptimizationSettings settings)
+    {
+        if (settings.DynamicAtrEnabled && settings.FlakyRetryEnabled == true)
+        {
+            var hasCustomBuckets = settings.DynamicAtrBuckets is not null
+                ? MetricTags.CIVisibilityDynamicAtrRetriesHasCustomBuckets.True
+                : MetricTags.CIVisibilityDynamicAtrRetriesHasCustomBuckets.False;
+            TelemetryFactory.Metrics.RecordCountCIVisibilityDynamicAtrRetries(hasCustomBuckets);
+        }
     }
 }