+++ tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/Testing/MsTestV2/TestMethodAttributeExecuteIntegration.cs
@@ -280,7 +280,9 @@ public sealed class TestMethodAttributeExecuteAsyncIntegration
 
                 // Flaky retry is enabled and the test failed
                 Interlocked.CompareExchange(ref _totalRetries, testOptimization.FlakyRetryFeature?.TotalFlakyRetryCount ?? TestOptimizationFlakyRetryFeature.TotalFlakyRetryCountDefault, -1);
-                var remainingRetries = testOptimization.FlakyRetryFeature?.FlakyRetryCount ?? TestOptimizationFlakyRetryFeature.FlakyRetryCountDefault;
+                var remainingRetries = testOptimization.Settings.DynamicAtrEnabled
+                    ? Common.GetDynamicAtrRetryCountForDuration(duration)
+                    : (testOptimization.FlakyRetryFeature?.FlakyRetryCount ?? TestOptimizationFlakyRetryFeature.FlakyRetryCountDefault);
                 if (remainingRetries > 0)
                 {
                     var retryState = new RetryState